requirements.txt
asgiref==3.8.1
certifi==2024.8.30
Django==4.2.16
django-unfold==0.39.0
gunicorn==23.0.0
packaging==24.0
pip==24.2
python-dotenv==1.0.1
sentry-sdk==2.15.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
